update test suite
