+++ .gitignore
@@ -6,3 +6,4 @@ coverage
 .env
 .ac-php-conf.json
 .phpunit.result.cache
+.DS_Store
